Conflicts:
	AddrMgr/AddrMgr.cpp
	AddrMgr/AddrMgr.h
	CHANGELOG
	ClntIfaceMgr/ClntIfaceMgr.h
	ClntMessages/ClntMsg.cpp
	ClntTransMgr/ClntTransMgr.cpp
	Options/Makefile.am
	Options/Makefile.in
	SrvTransMgr/SrvTransMgr.cpp
	SrvTransMgr/SrvTransMgr.h

+++ Options/OptReconfigureMsg.cpp
@@ -47,7 +47,7 @@ TOptReconfigureMsg::TOptReconfigureMsg(char *buf, int bufsize, TMsg* parent)
     }
 }
 
-int TOptReconfigureMsg::getSize()
+size_t TOptReconfigureMsg::getSize()
 {
     return 4 + 1; // header (4) + data-length (1)
 }
