+++ Options/OptUserClass.cpp
@@ -17,21 +17,17 @@
 #include "OptUserClass.h"
 
 TOptUserClass::TOptUserClass( char * &buf,  int &n, TMsg* parent)
-	:TOpt(OPTION_USER_CLASS, parent)
-{
+    :TOpt(OPTION_USER_CLASS, parent) {
 
 }
 
- int TOptUserClass::getSize()
-{
+size_t TOptUserClass::getSize() {
     return 0;
 }
 
- char * TOptUserClass::storeSelf( char* buf)
-{
+char * TOptUserClass::storeSelf( char* buf) {
     return 0;
 }
-bool TOptUserClass::isValid()
-{
+bool TOptUserClass::isValid() {
     return true;
 }
